Updated git ignore with yarn changes, added pages to project, isolated tailwind styles, separated project on several features
