+++ examples/with-create-react-app/package.json
@@ -22,6 +22,7 @@
     "react": "^17.0.2",
     "react-bootstrap": "^2.5.0",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.10.0",
     "react-scripts": "4.0.3",
     "react-table": "^7.8.0",
     "redux": "^4.2.0",
