+++ examples/with-create-react-app/src/feature/locations/data/request/getLocations.ts
@@ -6,7 +6,6 @@ import { LocationDto, LocationsDto } from '../dto/locations';
 export const GET_LOCATIONS_GQL = gql`
   query GetLocations {
     locations {
-      oi
       id
       name
       description
