+++ examples/with-create-react-app/src/feature/tailwind/ui/components/TailwindWidgetIframe/TailwindWidgetIframe.component.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const TailwindWidgetIframe: React.FC = () => {
+  return (
+    <>
+      <iframe frameBorder="0" title="Widget with tailwind" src="http://localhost:3000/tailwind" width={420} height={436} />
+    </>
+  );
+}
