+++ examples/with-create-react-app/src/feature/tailwind/ui/pages/tailwind.tsx
@@ -1,0 +1,5 @@
+import { AbstractColorsCard } from "../components/AbstractColorsCard/AbstractColorsCard.component"
+
+export const TailwindPage = () => {
+  return <div><AbstractColorsCard /></div>
+}
